Things that I have done
figured out that I was playing in the wrong place for buffering or loading the image.
	i have this confusing file that a folder is inside a folder and I'm supposed to put the pic in the folder -.-

changed the color of the red and blue goals to green so it can match the image. "figure out how to use color code" reminder for me*

added key pressed and key released stuff to WASD and ArrowKeys.